+++ CHANGELOG.md
@@ -4,13 +4,18 @@
 
 ### New Features
 - add lint for detecting duplicate features in capa-rules #2250 @v1bh475u
+- add span-of-calls scope to match features against a across a sliding window of API calls within a thread @williballenthin #2532
 - 
 ### Breaking Changes
 
-### New Rules (2)
+- add span-of-calls scope to rule format
+- capabilities functions return dataclasses instead of tuples
+
+### New Rules (3)
 
 - data-manipulation/encryption/rsa/encrypt-data-using-rsa-via-embedded-library Ana06
 - data-manipulation/encryption/use-bigint-function Ana06
+- nursery/dynamic-add-veh wballenthin@google.com
 -
 
 ### Bug Fixes